STORE-VUEX - The store module for Posts data has been setup - The History page has been split into components
