+++ src/store/Modules/postsModule/getters.js
@@ -1,0 +1,7 @@
+const getters = {
+  getPostsLength(state) {
+    return state.postsDataState.length;
+  },
+};
+
+export default getters;
