+++ src/store/Modules/postsModule/mutations.js
@@ -1,0 +1,7 @@
+const mutations = {
+  SET_POSTS_DATA_STATE(state, payload) {
+    state.postsDataState = payload;
+  },
+};
+
+export default mutations;
