+++ src/store/Modules/postsModule/state.js
@@ -1,0 +1,4 @@
+const state = {
+  postsDataState: [],
+};
+export default state;
